+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-## 0.5.0 (\_\_\_)
+## 0.5.1 (\_\_\_)
+
+**Fixes**:
+
+- **Controls**. _Need to add border by default otherwise it's invisible on the white body_
+- **CircleElement**. _Wrong tooltip position when segment is half of circle_
+
+## 0.5.0 (Mar 27, 2023)
 
 **BREAKING CHANGES**: